+++ src/__tests__/setup.ts
@@ -1,3 +1,4 @@
+
 jest.mock('../logger', () => {
   const original = jest.requireActual('../logger') as typeof import('../logger');
 
@@ -30,9 +31,9 @@ jest.mock('../config/awsConfig', () => ({
     sesAccessSecret: '98765',
     sesBounceAddress: 'bounce@example.com',
     sesBouncedEmailBucket: 'my-test-bucket',
-    dynamoTableName: 'test-table',
-    dynamoMaxQueryAttempts: 3,
-    dynamoEndpoint: 'http://localhost:8000',
+    sqs: {
+      generateMaDMPQueueUrl: 'http://sqs.example.com/queue/generateMadmp'
+    }
   }
 }));
 
@@ -99,5 +100,6 @@ jest.mock('../config/generalConfig', () => ({
     jwtRefreshSecret: 'testJwtRefreshSecret',
     jwtRefreshTTL: 500,
     hashTokenSecret: 'testTokenSecret',
-  }
+  },
+  envAsEnumValue: () => 'dev'
 }));